Update build to Gulp v4
